+++ src/utils/formatted/formatStatisticsData.ts
