+++ src/utils/index.ts
@@ -1,15 +1,17 @@
-export * from './calculateAverageMark'
-export * from './formatDate'
-export * from './extractMarksByDay'
-export * from './sortByDay'
-export * from './setDefaultMark'
-export * from './formatLessonDate'
-export * from './truncateString'
-export * from './logOut'
-export * from './isToday'
-export * from './handleResponse'
-export * from './formatStatisticsData'
-export * from './formatDateForRequest'
-export * from './textToLink'
+export * from './formatted/formatDate'
+export * from './formatted/formatDateForRequest'
+export * from './formatted/formatLessonDate'
+export * from './formatted/formatStatisticsData'
 export * from './getTimeRemaining'
-export * from './createSubjectMarksMap'
+export * from './handleResponse'
+export * from './isToday'
+export * from './logOut'
+export * from './marks/calculateAverageMark'
+export * from './marks/createSubjectMarksMap'
+export * from './marks/extractMarksByDay'
+export * from './marks/getBackgroundColor'
+export * from './marks/getSize'
+export * from './marks/setDefaultMark'
+export * from './sortByDay'
+export * from './textToLink'
+export * from './truncateString'
