+++ src/utils/marks/calculateAverageMark.ts
